consolidate function for crash logs
